+++ src/vega/packaging/parsers/pyproject.py
@@ -1,21 +1,23 @@
 """Module for holding the code for parsing the pyproject.toml files"""
 import os
+import re
 
 import toml
 
-from vega.packaging import commits
+from vega.packaging import commits, decorators
 from vega.packaging.parsers import abstract_parser
 
 
 class PyProject(abstract_parser.AbstractFileParser):
     """Parser for pyproject.toml files"""
-    FILENAME = "pyproject.toml"
+    FILENAME_REGEX = re.compile("pyproject.toml", re.I)
     TEMPLATE = {"build-system":
                     {"requires": ["setuptools >= 61.0"],
                      "build-backend": "setuptools.build_meta"},
                 "project": {
                     "name": None}}
     AUTOCREATE = False
+    PRIORITY = 1
 
     @property
     def version(self) -> str:
@@ -27,9 +29,7 @@ class PyProject(abstract_parser.AbstractFileParser):
     @property
     def content(self) -> dict:
         """The contents of this pyproject.toml file"""
-        if not self._content and self.exists:
-            self._content = toml.load(self.path)
-        return self._content or {}
+        return super(PyProject, self).content or {}
 
     def create(self):
         """Creates a pyproject.toml file with some default values."""
@@ -40,19 +40,17 @@ class PyProject(abstract_parser.AbstractFileParser):
         with open(self.path, "w+") as handle:
             toml.dump(content, handle)
 
+    def read(self) -> dict:
+        """Reads the contents of the pyproject.toml file"""
+        return toml.load(self.path)
+
+    @decorators.autocreate
     def update(self, commit_message: commits.CommitMessage):
         """Updates the contents of the pyproject.toml file with data from the commit message.
 
         Args:
             commit_message: the message to use for updating this file.
         """
-
-        # TODO: Create a decorator for this autocreate logic
-        if not self.exists and self.AUTOCREATE:
-            self.create()
-        elif not self.exists:
-            raise FileNotFoundError(f"{self.path} does not exist")
-
         self.update_version(commit_message)
 
         # Update pyproject.toml version